+++ src/components/element-hider/index.ts
@@ -1,1 +1,2 @@
+export * from './emote-holder-hider';
 export * from './top-gifters-hider';